requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
braceexpand==0.1.7
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
flake8==4.0.1
frozenlist==1.3.0
gitdb==4.0.9
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
import-expression==1.1.4
jishkucord==2.5.2
mccabe==0.6.1
multidict==6.0.2
protobuf==4.21.2
py-cord==2.0.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyNaCl==1.5.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
rfc3986==1.5.0
simpcalci==0.0.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
urllib3==1.26.9
yarl==1.7.2
youtube-dl==2021.12.17
